+++ stm32f103xx_i2c.h
@@ -91,4 +91,4 @@ void I2C_Init(I2C_Handle_t *pi2cHandler);
 void I2C_Control(I2C_TypeDef *pI2Cx , uint32_t EnorDi);
 void I2C_MasterSendData(I2C_Handle_t *pi2cHandler , uint8_t *pTxBuffer , uint32_t len ,uint8_t Slaveaddr);
 uint8_t I2C_GetFlagStatus(I2C_TypeDef *pI2Cx , uint8_t Flagname);
-void I2C_MasterReceiveData(I2C_Handle_t *pi2cHandler , uint16_t *pRxBuffer , uint32_t len ,uint8_t Slaveaddr);
+void I2C_MasterReceiveData(I2C_Handle_t *pi2cHandler , uint8_t *pRxBuffer , uint32_t len ,uint8_t Slaveaddr);
